Remove search id from each generation
We want to be able to record the trait-specs so we have added a new
`metadata` type that is passed to the record functions. It makes sense
to put the ID and the Search config in this as well, since we are
passing it anyway. Then it doesn't make any sense to include this in
each generation, since it will remain the same.